+++ cli/lm-cmd.py
@@ -573,6 +573,31 @@ class LM_CLI(GenericCLI, ControllerCLI):
                                 action='store_true',
                                 dest='_bars')
 
+        sp_watch = self.sp.add_parser('watch', help='Watch item state')
+        sp_watch.add_argument('i',
+                              help='Item ID (specify either ID or item type)',
+                              metavar='ID').completer = self.ComplLVAR(self)
+        sp_watch.add_argument('-r',
+                              '--interval',
+                              help='Watch interval (default: 1s)',
+                              metavar='SEC',
+                              default=1,
+                              type=float,
+                              dest='r')
+        sp_watch.add_argument('-n',
+                              '--rows',
+                              help='Rows to plot',
+                              metavar='NUM',
+                              type=int,
+                              dest='n')
+        sp_watch.add_argument('-x',
+                              '--prop',
+                              help='State prop to use (default: value)',
+                              choices=['status', 'value'],
+                              metavar='NUM',
+                              default='value',
+                              dest='x')
+
         sp_set = self.sp.add_parser('set', help='Set LVar state')
         sp_set.add_argument('i', help='LVar ID',
                             metavar='ID').completer = self.ComplLVAR(self)
@@ -1274,6 +1299,13 @@ class LM_CLI(GenericCLI, ControllerCLI):
             self.print_err('Macro code error: ' + str(e))
             return self.local_func_result_failed
 
+    def watch(self, props):
+        self.watch_item(props['i'],
+                        interval=props['r'],
+                        rows=props['n'],
+                        prop=props['x'])
+        return self.local_func_result_empty
+
 
 _me = 'EVA ICS LM CLI version %s' % __version__
 
@@ -1284,6 +1316,7 @@ if prog == 'eva-shell':
 cli = LM_CLI('lm', _me, prog=prog)
 
 _api_functions = {
+    'watch': cli.watch,
     'history': 'state_history',
     'config:get': 'get_config',
     'config:save': 'save_config',