+++ cli/uc-cmd.py
@@ -479,6 +479,31 @@ class UC_CLI(GenericCLI, ControllerCLI):
                                 action='store_true',
                                 dest='_bars')
 
+        sp_watch = self.sp.add_parser('watch', help='Watch item state')
+        sp_watch.add_argument('i',
+                              help='Item ID (specify either ID or item type)',
+                              metavar='ID').completer = self.ComplItemOID(self)
+        sp_watch.add_argument('-r',
+                              '--interval',
+                              help='Watch interval (default: 1s)',
+                              metavar='SEC',
+                              default=1,
+                              type=float,
+                              dest='r')
+        sp_watch.add_argument('-n',
+                              '--rows',
+                              help='Rows to plot',
+                              metavar='NUM',
+                              type=int,
+                              dest='n')
+        sp_watch.add_argument('-x',
+                              '--prop',
+                              help='State prop to use (default: value)',
+                              choices=['status', 'value'],
+                              metavar='NUM',
+                              default='value',
+                              dest='x')
+
         sp_update = self.sp.add_parser('update', help='Update item state')
         sp_update.add_argument('i', help='Item ID',
                                metavar='ID').completer = self.ComplItemOID(self)
@@ -1406,6 +1431,13 @@ class UC_CLI(GenericCLI, ControllerCLI):
             return self.local_func_result_failed
         return self.local_func_result_ok
 
+    def watch(self, props):
+        self.watch_item(props['i'],
+                        interval=props['r'],
+                        rows=props['n'],
+                        prop=props['x'])
+        return self.local_func_result_empty
+
 
 _me = 'EVA ICS UC CLI version %s' % __version__
 
@@ -1416,6 +1448,7 @@ if prog == 'eva-shell':
 cli = UC_CLI('uc', _me, prog=prog)
 
 _api_functions = {
+    'watch': cli.watch,
     'history': 'state_history',
     'action:exec': 'action',
     'action:result': 'result',